feat: vtable sheet context menu
