+++ packages/vtable-plugins/demo/vite.config.js
@@ -16,10 +16,7 @@ module.exports = {
       '@visactor/vtable/es': path.resolve(__dirname, '../../vtable/src/'),
       '@visactor/vtable': path.resolve(__dirname, '../../vtable/src/index.ts'),
       '@src': path.resolve(__dirname, '../../vtable/src/'),
-      '@vutils-extension': path.resolve(__dirname, '../../vtable/src/vutil-extension-temp'),
-      '@visactor/vrender-components': '/Users/bytedance/VisActor/VRender/packages/vrender-components/src/index.ts',
-      '@visactor/vrender-core': '/Users/bytedance/VisActor/VRender/packages/vrender-core/src/index.ts',
-      '@visactor/vrender-kits': '/Users/bytedance/VisActor/VRender/packages/vrender-kits/src/index.ts'
+      '@vutils-extension': path.resolve(__dirname, '../../vtable/src/vutil-extension-temp')
     }
   },
   plugins: [
