+++ packages/vtable-plugins/src/index.ts
@@ -14,3 +14,4 @@ export * from './gantt-export-image';
 export * from './wps-fill-handle';
 export * from './excel-import';
 export * from './table-series-number';
+export * from './context-menu';
