+++ packages/vtable-sheet/src/components/VTableSheet.ts
@@ -565,6 +565,7 @@ export default class VTableSheet {
       defaultColWidth: this.options.defaultColWidth,
       parent: this,
       plugins: getTablePlugins(),
+      headerEditor: 'input',
       editor: 'input',
       select: {
         makeSelectCellVisible: false
