+++ packages/vtable-sheet/src/core/table-plugins.ts
@@ -46,6 +46,7 @@ export function getTablePlugins() {
     colHighlight: true,
     rowHighlight: true
   });
+  const contextMenuPlugin = new VTablePlugins.ContextMenuPlugin({});
   const excelEditCellKeyboardPlugin = new VTablePlugins.ExcelEditCellKeyboardPlugin();
-  return [addRowColumn, highlightPlugin, excelEditCellKeyboardPlugin, tableSeriesNumberPlugin];
+  return [addRowColumn, highlightPlugin, excelEditCellKeyboardPlugin, tableSeriesNumberPlugin, contextMenuPlugin];
 }
