+++ packages/vtable/package.json
@@ -52,10 +52,10 @@
   },
   "dependencies": {
     "@visactor/vtable-editors": "workspace:*",
-    "@visactor/vrender-core": "1.0.8-alpha.2",
-    "@visactor/vrender-kits": "1.0.8-alpha.2",
-    "@visactor/vrender-components": "1.0.8-alpha.2",
-    "@visactor/vrender-animate": "1.0.8-alpha.2",
+    "@visactor/vrender-core": "1.0.9-alpha.2",
+    "@visactor/vrender-kits": "1.0.9-alpha.2",
+    "@visactor/vrender-components": "1.0.9-alpha.2",
+    "@visactor/vrender-animate": "1.0.9-alpha.2",
     "@visactor/vutils": "~0.19.1",
     "@visactor/vscale": "~0.18.1",
     "@visactor/vdataset": "~0.18.1",
