+++ packages/vtable/src/tools/style.ts
@@ -1,5 +1,5 @@
 import { isArray, isValid } from '@visactor/vutils';
-import type { ScrollStyle } from '../ts-types';
+import type { BaseTableAPI, ScrollStyle } from '../ts-types';
 
 export function getHorizontalScrollBarSize(scrollStyle?: ScrollStyle): number {
   if (
@@ -36,4 +36,22 @@ export function isValidStyle(style: (string | number) | (string | number)[]) {
 
 export function isZeroStyle(style: number | number[]) {
   return style === 0 || (isArray(style) && style.every(s => s === 0));
+}
+
+// 设置复制区域的状态 类似excel的虚线框
+export function setActiveCellRangeState(table: BaseTableAPI) {
+  const selectRanges = table.stateManager.select.ranges;
+  const setRanges = [];
+  for (let i = 0; i < selectRanges.length; i++) {
+    const range = selectRanges[i];
+    setRanges.push({
+      range,
+      style: {
+        cellBorderColor: 'blue',
+        cellBorderLineWidth: 2,
+        cellBorderLineDash: [5, 5]
+      }
+    });
+  }
+  table.stateManager.setCustomSelectRanges(setRanges);
 }
